New file for button logic;
Finished demo template;

Reconstruct WaveData: now duration will be derived by parameters through the function defined in device class, and will not be storaged by a single varible with specific value that can not be changed once the duration is registered

With the reconstruction of WaveData, other class used with WaveData to support the function has been modified to adapt the change

Add HTMLGen. and CheckSelect. class to support new function
1. New log interface
2. Read/Write device to the specific file to persist data